GET /money_records

+++ backend/domain/money.go
@@ -3,12 +3,12 @@ package domain
 type Money struct {
 	ID int32
 	// DO:ある程度記録が溜まったら、最初の方を削除？
-	TypeID int64
+	TypeID int32
 	// 収支を行うUserID
 	UserID int64
 	Amount int64
 	// User1を基準とした精算額
-	CalculationUser1 int64
+	CalculationUser1 float64
 	CreatedAt        string
 	// UpdatedAt        string DO:フロントエンドから修正できるようにする？
 }
